Pull out functions using model and/or state, fix build in progress
