+++ src/testing/mod.rs
@@ -199,10 +199,7 @@ impl<'a, L: Ledger, N: MockNetwork<'a, L>> MockLedger<'a, L, N> {
 // that cannot directly be compared for equality. It is sufficient for tests that want to compare
 // keystore states (like round-trip serialization tests) but since it is deterministic, we shouldn't
 // make it into a PartialEq instance.
-pub fn assert_keystore_states_eq<'a, L: Ledger>(
-    w1: &LedgerState<'a, L>,
-    w2: &LedgerState<'a, L>,
-) {
+pub fn assert_keystore_states_eq<'a, L: Ledger>(w1: &LedgerState<'a, L>, w2: &LedgerState<'a, L>) {
     assert_eq!(w1.txn_state.now(), w2.txn_state.now());
     assert_eq!(
         w1.txn_state.validator.commit(),
@@ -211,8 +208,8 @@ pub fn assert_keystore_states_eq<'a, L: Ledger>(
     assert_eq!(w1.proving_keys, w2.proving_keys);
     assert_eq!(w1.txn_state.nullifiers, w2.txn_state.nullifiers);
     assert_eq!(
-        w1.txn_state.record_mt.commitment(),
-        w2.txn_state.record_mt.commitment()
+        w1.txn_state.record_mt().commitment(),
+        w2.txn_state.record_mt().commitment()
     );
 }
 
@@ -527,10 +524,10 @@ pub trait SystemUnderTest<'a>: Default + Send + Sync {
         )],
     ) {
         for (keystore, _, _) in keystores {
-            let KeystoreSharedState { state, model, .. } = &*keystore.mutex.read().await;
-            let persistence = &model.persistence;
+            let KeystoreSharedState { model, .. } = &*keystore.mutex.read().await;
+            let state = &model.ledger_state;
             let state = state.clone();
-            let loaded = persistence.load().await.unwrap();
+            let loaded = state.load().await.unwrap();
             assert_keystore_states_eq(&state, &loaded);
         }
     }
